+++ include/bloom_filter/cbf.hpp
@@ -10,7 +10,7 @@
 
 class CountingBloomFilter {
    public:
-    CountingBloomFilter(size_t m, uint8_t k);
+    CountingBloomFilter(uint64_t m, uint8_t k);
     void insert(uint8_t *key, uint16_t keyLength);
     bool lookup(uint8_t *key, uint16_t keyLength);
     bool remove(uint8_t *key, uint16_t keyLength);
