+++ src/bloom_filter/cbf.cpp
@@ -1,6 +1,6 @@
 #include "cbf.hpp"
 
-CountingBloomFilter::CountingBloomFilter(size_t m, uint8_t k): 
+CountingBloomFilter::CountingBloomFilter(uint64_t m, uint8_t k): 
     f_set{setOfMurmurHash64A(k)},
     filter(roundUp8(m) >> 3),
     slots{roundUp8(m) >> 2}
